Up the warnings in ipc (take 2) - add chromium_code:1 to the GYP file - Fix some unittest compares of literal 0 to apis that return size_t - initializer order match declared order - type_id is a uint32, so fix up comparison warnings by using the right type in the test code. - duplicate a type cast used in the ipc headers into the ipc impl to make windows happy. - msvc warns about getenv, avoid it.
BUG=none
TEST=everything still builds/works
Review URL: http://codereview.chromium.org/2821028

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@53468 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7ee1a44c27384650612290a18ccbe736e0e4b955

+++ ipc/ipc_message_utils.cc
@@ -55,7 +55,7 @@ static void WriteValue(Message* m, const Value* value, int recursion) {
     }
     case Value::TYPE_BINARY: {
       const BinaryValue* binary = static_cast<const BinaryValue*>(value);
-      m->WriteData(binary->GetBuffer(), binary->GetSize());
+      m->WriteData(binary->GetBuffer(), static_cast<int>(binary->GetSize()));
       break;
     }
     case Value::TYPE_DICTIONARY: {
